Split the orchestrator along its seams; move pure Three.js code to the three/ layer
room-organizer.tsx (1156 lines) contained three self-contained clusters
that communicated only via refs and context — extracted verbatim:

- hooks/use-item-drag.ts: the drag fast-path (session refs, group drag,
  collision tint, camera wall-seat on release)
- hooks/use-item-placement.ts: snapPosition, getDragPlaneY, and the
  wall-aware placeCatalogItem
- hooks/use-import-export.ts: screenshot, GLB export, share link, and
  JSON import handlers

use-three-scene.ts embedded two plain non-React functions that belong
in the three/ layer per the module rules:

- attachDragHandlers (~155 lines of pointer/raycast logic) moves to
  three/drag-handlers.ts, which now owns the HoverInfo/SelectionMode/
  SceneEventHandlers types (re-exported from the hook for compatibility)
- addLights moves into three/lighting.ts, which already owned the
  light:* tags it sets and applyTimeOfDay consumes; the hemisphere tag
  joins LIGHTING_TAGS

Net: room-organizer.tsx 1156 -> 854 lines, use-three-scene.ts 535 ->
304. Pure code motion — verified in the running app (select, unlock,
mouse-drag commits the new position and re-locks; zero console errors).

Fixes #37
