+++ components/room-organizer/hooks/use-three-scene.ts
@@ -1,37 +1,17 @@
 import { useEffect, useRef, useState } from 'react';
+import { attachDragHandlers } from '../three/drag-handlers';
+import { addLights } from '../three/lighting';
 import { setMaxAnisotropy } from '../three/texture-settings';
+import type { SceneEventHandlers } from '../three/drag-handlers';
 import type * as ThreeNS from 'three';
 import type { OrbitControls as OrbitControlsType } from 'three/examples/jsm/controls/OrbitControls.js';
 
 type ThreeModule = typeof import('three');
 
-export interface HoverInfo {
-  id: string;
-  clientX: number;
-  clientY: number;
-}
+export type { HoverInfo, SelectionMode } from '../three/drag-handlers';
 
-export type SelectionMode = 'replace' | 'toggle';
-
-export interface UseThreeSceneOptions {
+export interface UseThreeSceneOptions extends SceneEventHandlers {
   canvasRef: React.RefObject<HTMLCanvasElement>;
-  onItemSelect: (id: string, mode: SelectionMode) => void;
-  onItemDragStart?: (id: string) => void;
-  onItemDrag: (id: string, x: number, z: number) => void;
-  onItemDragEnd?: (id: string) => void;
-  onItemHover?: (info: HoverInfo | null) => void;
-  /** Fired when the user clicks an empty area (no furniture hit). */
-  onEmptyClick?: (x: number, z: number) => void;
-  /** Fired when the user clicks a wall (exterior or interior). */
-  onWallSelect?: (info: { wallId: string; kind: 'exterior' | 'interior' }) => void;
-  /** Fired on every mousemove that lands on the floor (no drag in progress). */
-  onFloorPointerMove?: (x: number, z: number) => void;
-  /** Fired when the pointer leaves the canvas, so floor-tracking UI can clean up. */
-  onFloorPointerLeave?: () => void;
-  /** Adjust a candidate drop position before it's applied (snap-to-grid, snap-to-wall, etc.). */
-  snapPosition: (itemId: string, x: number, z: number) => { x: number; z: number };
-  /** World-space Y for the drag plane (the floor the user is editing). */
-  getDragPlaneY?: () => number;
 }
 
 export interface UseThreeSceneResult {
@@ -315,218 +295,7 @@ function supportsWebGL(): boolean {
   }
 }
 
-function addLights(THREE: ThreeModule, scene: ThreeNS.Scene): void {
-  // Hemisphere fill gives the bright sky / warm ground bounce a suburban lot
-  // reads with — without it everything in shadow goes flat-grey.
-  const hemi = new THREE.HemisphereLight(0xbfe5ff, 0xa07a48, 0.55);
-  hemi.userData.type = 'light:hemi';
-  scene.add(hemi);
 
-  const ambient = new THREE.AmbientLight(0xffffff, 0.45);
-  ambient.userData.type = 'light:ambient';
-  scene.add(ambient);
-
-  const directional = new THREE.DirectionalLight(0xfff4d1, 1.05);
-  directional.position.set(7, 14, 6);
-  directional.castShadow = true;
-  // Frustum large enough to cover the lot + the outdoor perimeter so trees
-  // and the room walls all cast contact shadows on the grass.
-  const shadowExtent = 24;
-  directional.shadow.camera.left = -shadowExtent;
-  directional.shadow.camera.right = shadowExtent;
-  directional.shadow.camera.top = shadowExtent;
-  directional.shadow.camera.bottom = -shadowExtent;
-  directional.shadow.camera.near = 1;
-  directional.shadow.camera.far = 60;
-  directional.shadow.mapSize.set(2048, 2048);
-  directional.shadow.bias = -0.0005;
-  directional.userData.type = 'light:directional';
-  scene.add(directional);
-}
-
-interface DragHandlersOptions {
-  THREE: ThreeModule;
-  canvas: HTMLCanvasElement;
-  camera: ThreeNS.PerspectiveCamera;
-  scene: ThreeNS.Scene;
-  controls: OrbitControlsType;
-  markDirty: () => void;
-  handlersRef: React.MutableRefObject<{
-    onItemSelect: (id: string, mode: SelectionMode) => void;
-    onItemDragStart?: (id: string) => void;
-    onItemDrag: (id: string, x: number, z: number) => void;
-    onItemDragEnd?: (id: string) => void;
-    onItemHover?: (info: HoverInfo | null) => void;
-    onEmptyClick?: (x: number, z: number) => void;
-    onWallSelect?: (info: { wallId: string; kind: 'exterior' | 'interior' }) => void;
-    onFloorPointerMove?: (x: number, z: number) => void;
-    onFloorPointerLeave?: () => void;
-    snapPosition: (id: string, x: number, z: number) => { x: number; z: number };
-    getDragPlaneY?: () => number;
-  }>;
-}
-
-function attachDragHandlers({
-  THREE,
-  canvas,
-  camera,
-  scene,
-  controls,
-  markDirty,
-  handlersRef,
-}: DragHandlersOptions): () => void {
-  const raycaster = new THREE.Raycaster();
-  const pointer = new THREE.Vector2();
-  const dragPlane = new THREE.Plane(new THREE.Vector3(0, 1, 0), 0);
-  const intersection = new THREE.Vector3();
-
-  let dragTarget: ThreeNS.Object3D | null = null;
-
-  const setPointerFromEvent = (event: MouseEvent): void => {
-    const rect = canvas.getBoundingClientRect();
-    pointer.x = ((event.clientX - rect.left) / rect.width) * 2 - 1;
-    pointer.y = -((event.clientY - rect.top) / rect.height) * 2 + 1;
-  };
-
-  const onMouseDown = (event: MouseEvent): void => {
-    setPointerFromEvent(event);
-    raycaster.setFromCamera(pointer, camera);
-    const furniture = scene.children.filter((obj) => obj.userData.type === 'furniture');
-    const hits = raycaster.intersectObjects(furniture, true);
-    if (hits.length === 0) {
-      // Walls get a chance to claim the click before we fall through to the
-      // floor's onEmptyClick handler. Required for "pick a wall to
-      // paint it" interaction.
-      if (handlersRef.current.onWallSelect) {
-        const wallObjects = scene.children.filter(
-          (obj) =>
-            obj.userData.type === 'wall' || obj.userData.type === 'interior-wall'
-        );
-        const wallHits = raycaster.intersectObjects(wallObjects, false);
-        const wallHit = wallHits.find((h) => h.object.visible);
-        if (wallHit) {
-          const id = wallHit.object.userData.wallId as string | undefined;
-          if (id) {
-            handlersRef.current.onWallSelect({
-              wallId: id,
-              kind: wallHit.object.userData.type === 'interior-wall'
-                ? 'interior'
-                : 'exterior',
-            });
-            return;
-          }
-        }
-      }
-      if (handlersRef.current.onEmptyClick) {
-        dragPlane.constant = -(handlersRef.current.getDragPlaneY?.() ?? 0);
-        if (raycaster.ray.intersectPlane(dragPlane, intersection)) {
-          handlersRef.current.onEmptyClick(intersection.x, intersection.z);
-        }
-      }
-      return;
-    }
-
-    const target = ascendToFurniture(hits[0]?.object);
-    if (!target) return;
-
-    const mode: SelectionMode = event.ctrlKey || event.metaKey ? 'toggle' : 'replace';
-    const itemId = target.userData.id as string;
-    handlersRef.current.onItemSelect(itemId, mode);
-
-    if (target.userData.locked === true || mode === 'toggle') return;
-
-    dragTarget = target;
-    controls.enabled = false;
-    handlersRef.current.onItemDragStart?.(itemId);
-  };
-
-  let lastHoverId: string | null = null;
-  const updateHover = (event: MouseEvent): void => {
-    const hoverCallback = handlersRef.current.onItemHover;
-    if (!hoverCallback) return;
-    setPointerFromEvent(event);
-    raycaster.setFromCamera(pointer, camera);
-    const furniture = scene.children.filter((obj) => obj.userData.type === 'furniture');
-    const hits = raycaster.intersectObjects(furniture, true);
-    const target = ascendToFurniture(hits[0]?.object);
-    const id = target ? (target.userData.id as string) : null;
-
-    // Only notify on hover changes — a fresh payload per mousemove would
-    // re-render React for every pixel of travel. The tooltip anchors at the
-    // point where the hover began.
-    if (id === lastHoverId) return;
-    lastHoverId = id;
-    canvas.style.cursor = id ? 'pointer' : '';
-    hoverCallback(id ? { id, clientX: event.clientX, clientY: event.clientY } : null);
-  };
-
-  const onMouseMove = (event: MouseEvent): void => {
-    if (!dragTarget) {
-      updateHover(event);
-      if (handlersRef.current.onFloorPointerMove) {
-        setPointerFromEvent(event);
-        raycaster.setFromCamera(pointer, camera);
-        dragPlane.constant = -(handlersRef.current.getDragPlaneY?.() ?? 0);
-        if (raycaster.ray.intersectPlane(dragPlane, intersection)) {
-          handlersRef.current.onFloorPointerMove(intersection.x, intersection.z);
-        }
-      }
-      return;
-    }
-    setPointerFromEvent(event);
-    raycaster.setFromCamera(pointer, camera);
-    dragPlane.constant = -(handlersRef.current.getDragPlaneY?.() ?? 0);
-    raycaster.ray.intersectPlane(dragPlane, intersection);
-
-    const itemId = dragTarget.userData.id as string;
-    const snapped = handlersRef.current.snapPosition(itemId, intersection.x, intersection.z);
-
-    dragTarget.position.x = snapped.x;
-    dragTarget.position.z = snapped.z;
-    markDirty();
-    handlersRef.current.onItemDrag(itemId, snapped.x, snapped.z);
-  };
-
-  const onMouseUp = (): void => {
-    if (!dragTarget) return;
-    const id = dragTarget.userData.id as string;
-    dragTarget = null;
-    controls.enabled = true;
-    handlersRef.current.onItemDragEnd?.(id);
-  };
-
-  const onMouseLeave = (): void => {
-    handlersRef.current.onItemHover?.(null);
-    handlersRef.current.onFloorPointerLeave?.();
-    canvas.style.cursor = '';
-    if (dragTarget) {
-      const id = dragTarget.userData.id as string;
-      dragTarget = null;
-      controls.enabled = true;
-      handlersRef.current.onItemDragEnd?.(id);
-    }
-  };
-
-  canvas.addEventListener('mousedown', onMouseDown);
-  canvas.addEventListener('mousemove', onMouseMove);
-  canvas.addEventListener('mouseup', onMouseUp);
-  canvas.addEventListener('mouseleave', onMouseLeave);
-
-  return () => {
-    canvas.removeEventListener('mousedown', onMouseDown);
-    canvas.removeEventListener('mousemove', onMouseMove);
-    canvas.removeEventListener('mouseup', onMouseUp);
-    canvas.removeEventListener('mouseleave', onMouseLeave);
-  };
-}
-
-function ascendToFurniture(node: ThreeNS.Object3D | undefined): ThreeNS.Object3D | null {
-  let current: ThreeNS.Object3D | null = node ?? null;
-  while (current && current.userData?.type !== 'furniture') {
-    current = current.parent;
-  }
-  return current && current.userData?.type === 'furniture' ? current : null;
-}
 
 function messageOf(error: unknown, fallback: string): string {
   if (error instanceof Error) return error.message;
